+++ qmail-smtpd.c
@@ -28,9 +28,9 @@
 unsigned int databytes = 0;
 int timeout = 1200;
 
-int safewrite(fd,buf,len) int fd; char *buf; int len;
+ssize_t safewrite(fd,buf,len) int fd; char *buf; int len;
 {
-  int r;
+  ssize_t r;
   r = timeoutwrite(timeout,fd,buf,len);
   if (r <= 0) _exit(1);
   return r;
@@ -265,9 +265,9 @@ void smtp_rcpt(arg) char *arg; {
 }
 
 
-int saferead(fd,buf,len) int fd; char *buf; int len;
+ssize_t saferead(fd,buf,len) int fd; char *buf; int len;
 {
-  int r;
+  ssize_t r;
   flush();
   r = timeoutread(timeout,fd,buf,len);
   if (r == -1) if (errno == error_timeout) die_alarm();
